Max Zoom of layer on map type change
This little edit checks if the current leaflet zoom is too high for the current Google layer. This is especially a problem with the Terrain layer.

Now, when you switch to the terrain layer with a very high zoom level, the map zooms out to the maximum google terrain level.

This prevents that your overlays are on the wrong zoom level compared to the google layer. However, if you zoom in past the max zoom level on Terrain, the check does not work. It only works if you switch to Terrain. maybe someone can figure out how to always make the check work.